+++ NEWS.md
@@ -1,6 +1,8 @@
 # chromatographR 0.3.1
 
 * Added support for variable penalty dynamic time warping in `correct_RT`.
+* Allow use of raw data for peak integration in `get_peaks`.
+* Added `verbose` option to `correct_rt` to print reference chromatogram.
 
 # chromatographR 0.3